fixed all the signed/unsigned mismatch warnings
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@2591 d7cf8633-e32d-0410-b094-e92efae38249

+++ common.c
@@ -375,7 +375,7 @@ char *MSG_ReadString (void)
 			break;
 		string[l] = c;
 		l++;
-	} while (l < sizeof(string)-1);
+	} while (l < (int)sizeof(string)-1);
 
 	string[l] = 0;
 
@@ -1083,7 +1083,7 @@ void COM_CopyFile (char *netpath, char *cachepath)
 
 	while (remaining)
 	{
-		if (remaining < sizeof(buf))
+		if (remaining < (int)sizeof(buf))
 			count = remaining;
 		else
 			count = sizeof(buf);
